:bulb: Home - networking 보정 작업
- model 값 내에 비중을 나타내는 값 자료형 수정 (Int -> Double)

+++ Projects/Home/Targets/HomeKit/Sources/Network/ReponseModel/QuizInfoModel.swift
@@ -44,13 +44,13 @@ public struct AnswerModel: Decodable, Identifiable {
     public var optionId: Int
     public var content: String
     public var isCorrect: Bool
-    public var selectivity: Int
+    public var selectivity: Double
     
     public var id: Int {
         return optionId
     }
     
-    public init(optionId: Int, content: String, isCorrect: Bool, selectivity: Int) {
+    public init(optionId: Int, content: String, isCorrect: Bool, selectivity: Double) {
         self.optionId = optionId
         self.content = content
         self.isCorrect = isCorrect
